refactor(cli): move the run path into src/run.ts
A pure move, no behaviour change. src/index.ts is 20 static imports and two
blocks of top-level statements, so a dispatcher cannot live at the top of it:
the adapter registry, the exporter registry and the mqtt-proxy graph are all
built at module evaluation, before any argument is read. Splitting the move
from the dispatcher itself keeps the next diff readable.

Refs #365
